Rename is_palindromic to is_palindrome
