+++ libs/algorithm/doc/algorithm.qbk
@@ -49,7 +49,7 @@ Thanks to all the people who have reviewed this library and made suggestions for
 [include one_of.qbk]
 [include is_partitioned.qbk]
 [include partition_point.qbk]
-[include is_palindromic.qbk]
+[include is_palindrome.qbk]
 
 [xinclude autodoc.xml]
