+++ libs/algorithm/test/Jamfile.v2
@@ -64,4 +64,4 @@ run partition_copy_test1.cpp ;
 
 run find_if_not_test1.cpp ;
 
-run is_palindromic_test.cpp ;
+run is_palindrome_test.cpp ;
